More work on improved version. Added additional vocab + Blip + GitVision implementation
