+++ .gitignore
@@ -5,3 +5,5 @@
 /demo_img.png
 /astronaut_with_beer.jpg
 /monkey_with_gun.jpg
+/English-Morph.txt
+/fruit_bowl.jpg
