+++ image_captioning.py
@@ -3,7 +3,6 @@ import os
 import requests
 import clip
 import cv2
-import matplotlib.pyplot as plt
 import numpy as np
 from PIL import Image
 from profanity_filter import ProfanityFilter
@@ -14,16 +13,24 @@ from utils import print_time_dec
 
 class ImageManager:
     def __init__(self):
+        self.images_to_download = {
+            'demo_img.png': 'https://github.com/rmokady/CLIP_prefix_caption/raw/main/Images/COCO_val2014_000000165547.jpg',
+            'monkey_with_gun.jpg': 'https://drive.google.com/uc?export=download&id=1iG0TJTZ0yRJEC8dA-WwS7X-GhuX8sfy8',
+            'astronaut_with_beer.jpg': 'https://drive.google.com/uc?export=download&id=1p5RwifMFtl1CLlUXaIR_y60_laDTbNMi',
+            'fruit_bowl.jpg': 'https://drive.google.com/uc?export=download&id=1gRYMoTfCwuV4tNy14Qf2Q_hebx05GNd9'
+        }
         self.download_data()
 
+    def download_data(self):
+        # Download images
+        for img_path, img_url in self.images_to_download.items():
+            if not os.path.exists(img_path):
+                self.download_image_from_url(img_path, img_url)
+
     @staticmethod
-    def download_data():
-        # Download test image
-        fname = 'demo_img.png'
-        if not os.path.exists(fname):
-            img_url = "https://github.com/rmokady/CLIP_prefix_caption/raw/main/Images/COCO_val2014_000000165547.jpg"
-            with open(fname, 'wb') as f:
-                f.write(requests.get(img_url).content)
+    def download_image_from_url(img_path, img_url):
+        with open(img_path, 'wb') as f:
+            f.write(requests.get(img_url).content)
 
     @staticmethod
     def load_image(image_path):
@@ -46,7 +53,8 @@ class VocabManager:
                 f.write(response.content)
         # Download object categories from Tencent ML Images.
         if not os.path.exists('dictionary_and_semantic_hierarchy.txt'):
-            url = "https://raw.githubusercontent.com/Tencent/tencent-ml-images/master/data/dictionary_and_semantic_hierarchy.txt"
+            url = "https://raw.githubusercontent.com/Tencent/tencent-ml-images/master/data/dictionary_and_semantic_hier"
+            "archy.txt"
             response = requests.get(url)
             with open("dictionary_and_semantic_hierarchy.txt", "wb") as f:
                 f.write(response.content)
@@ -137,16 +145,33 @@ class ClipManager:
         high_to_low_scores = np.sort(scores).squeeze()[::-1]
         return high_to_low_texts, high_to_low_scores
 
+    def get_image_caption_score(self, caption, img_feats):
+        text_feats = self.get_text_feats([caption])
+        return float(text_feats @ img_feats.T)
+
 
 class FlanT5Manager:
     def __init__(self, version="google/flan-t5-xl"):
+        # Instantiate the model
         self.model = AutoModelForSeq2SeqLM.from_pretrained(version)
+        # Instantiate the tokenizer
         self.tokenizer = AutoTokenizer.from_pretrained(version)
 
+    @print_time_dec
     def generate_response(self, prompt, model_params=None):
+        """
+        :param prompt: Prompt(s) as list or str
+        :param model_params: Model parameters
+        :return: str if 1 else list
+        """
+        if model_params is None:
+            model_params = {}
         inputs = self.tokenizer(prompt, return_tensors="pt")
         outputs = self.model.generate(**inputs, **model_params)
-        return self.tokenizer.batch_decode(outputs, skip_special_tokens=True)[0]
+        decoded = self.tokenizer.batch_decode(outputs, skip_special_tokens=True)
+        if len(decoded) == 1:
+            return decoded[0]
+        return decoded
 
 
 def num_params(model):
@@ -160,113 +185,3 @@ def print_clip_info(model):
     print("Input image resolution:", model.visual.input_resolution)
     print("Context length:", model.context_length)
     print("Vocab size:", model.vocab_size)
-
-
-def main(img_path='demo_img.png', verbose=True):
-
-    # Set the device to use
-    if getattr(torch, 'has_mps', False):
-        device = 'mps'
-    elif torch.cuda.is_available():
-        device = 'gpu'
-    else:
-        device = 'cpu'
-
-    # Instantiate the clip manager
-    clip_manager = ClipManager(device)
-
-    # Instantiate the image manager
-    image_manager = ImageManager()
-
-    # Instantiate the vocab manager
-    vocab_manager = VocabManager()
-
-    # Instantiate the Flan T5 manager
-    flan_manager = FlanT5Manager()
-
-    # Print out clip model info
-    print_clip_info(clip_manager.model)
-
-    # Calculate the place features
-    place_feats = clip_manager.get_text_feats([f'Photo of a {p}.' for p in vocab_manager.place_list])
-
-    # Calculate the object features
-    object_feats = clip_manager.get_text_feats([f'Photo of a {o}.' for o in vocab_manager.object_list])
-
-    # Load image.
-    img = image_manager.load_image(img_path)
-    img_feats = clip_manager.get_img_feats(img)
-    plt.imshow(img)
-    plt.show()
-
-    # Zero-shot VLM: classify image type.
-    img_types = ['photo', 'cartoon', 'sketch', 'painting']
-    img_types_feats = clip_manager.get_text_feats([f'This is a {t}.' for t in img_types])
-    sorted_img_types, img_type_scores = clip_manager.get_nn_text(img_types, img_types_feats, img_feats)
-    img_type = sorted_img_types[0]
-
-    # Zero-shot VLM: classify number of people.
-    ppl_texts = ['no people', 'people']
-    ppl_feats = clip_manager.get_text_feats([f'There are {p} in this photo.' for p in ppl_texts])
-    sorted_ppl_texts, ppl_scores = clip_manager.get_nn_text(ppl_texts, ppl_feats, img_feats)
-    ppl_result = sorted_ppl_texts[0]
-    if ppl_result == 'people':
-        ppl_texts = ['is one person', 'are two people', 'are three people', 'are several people', 'are many people']
-        ppl_feats = clip_manager.get_text_feats([f'There {p} in this photo.' for p in ppl_texts])
-        sorted_ppl_texts, ppl_scores = clip_manager.get_nn_text(ppl_texts, ppl_feats, img_feats)
-        ppl_result = sorted_ppl_texts[0]
-    else:
-        ppl_result = f'are {ppl_result}'
-
-    # Zero-shot VLM: classify places.
-    place_topk = 3
-    sorted_places, places_scores = clip_manager.get_nn_text(vocab_manager.place_list, place_feats, img_feats)
-
-    # Zero-shot VLM: classify objects.
-    obj_topk = 10
-    sorted_obj_texts, obj_scores = clip_manager.get_nn_text(vocab_manager.object_list, object_feats, img_feats)
-    object_list = ''
-    for i in range(obj_topk):
-        object_list += f'{sorted_obj_texts[i]}, '
-    object_list = object_list[:-2]
-
-    # Zero-shot LM: generate captions.
-    num_captions = 10
-    prompt = f'''I am an intelligent image captioning bot.
-    This image is a {img_type}. There {ppl_result}.
-    I think this photo was taken at a {sorted_places[0]}, {sorted_places[1]}, or {sorted_places[2]}.
-    I think there might be a {object_list} in this {img_type}.
-    A creative short caption I can generate to describe this image is:'''
-
-    model_params = {'temperature': 0.9, 'max_length': 40, 'do_sample': True}
-    caption_texts = [flan_manager.generate_response(prompt, model_params) for _ in range(num_captions)]
-
-    # Zero-shot VLM: rank captions.
-    caption_feats = clip_manager.get_text_feats(caption_texts)
-    sorted_captions, caption_scores = clip_manager.get_nn_text(caption_texts, caption_feats, img_feats)
-    print(f'{sorted_captions[0]}\n')
-
-    if verbose:
-        print(f'VLM: This image is a:')
-        for img_type, score in zip(sorted_img_types, img_type_scores):
-            print(f'{score:.4f} {img_type}')
-
-        print(f'\nVLM: There:')
-        for ppl_text, score in zip(sorted_ppl_texts, ppl_scores):
-            print(f'{score:.4f} {ppl_text}')
-
-        print(f'\nVLM: I think this photo was taken at a:')
-        for place, score in zip(sorted_places[:place_topk], places_scores[:place_topk]):
-            print(f'{score:.4f} {place}')
-
-        print(f'\nVLM: I think there might be a:')
-        for obj_text, score in zip(sorted_obj_texts[:obj_topk], obj_scores[:obj_topk]):
-            print(f'{score:.4f} {obj_text}')
-
-        print(f'\nLM generated captions ranked by VLM scores:')
-        for caption, score in zip(sorted_captions, caption_scores):
-            print(f'{score:.4f} {caption}')
-
-
-if __name__ == '__main__':
-    main()
